googlegenaiproject/requirements.txt
google-genai>=1.2.0
google-cloud-firestore>=2.16.0
google-cloud-secret-manager>=2.20.0
google-auth>=2.29.0
pydantic>=2.9.0
python-dotenv>=1.0.1
rich>=13.7.1
tenacity>=9.0.0
httpx>=0.27.2
requests>=2.32.3
pytest>=8.3.3
pypdf>=4.3.1
python-docx>=1.1.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
chromadb==1.3.0
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
crewai==0.193.2
cryptography==46.0.3
debugpy==1.8.16
decorator==5.2.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.120.4
filelock==3.19.1
Flask==3.1.2
Flask-Cors==4.0.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.68.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
identify==2.6.15
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.12.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
json_repair==0.25.2
jsonpickle==4.1.1
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kubernetes==33.1.0
litellm==1.74.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.109.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.4.0
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.3.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.5
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyvis==0.3.2
python-multipart
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.49.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.0
tomli==2.3.0
tomli_w==1.2.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uv==0.9.7
uvicorn==0.38.0
vertexai==1.68.0
virtualenv==20.35.4
watchfiles==1.1.1
wcwidth==0.2.13
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
whisper==1.1.10
yarl==1.22.0
zipp==3.23.0
